Hab durchgezogen und die Game Engine für den Push und use von den anderen gepolished. Movement funktioniert jetzt besser, hat vorher gestottert, Enemies und die Map werden außerhalb vom Screen nicht mehr gerendert um Performance zu sparen. Außer dem funktioniert meine Serialisation mit Save UND Load von den Spieler Daten (ENDLLICH!!!!)

+++ src/main/java/GUI/Camera.java
@@ -3,8 +3,8 @@ package GUI;
 import Player.Player;
 
 public class Camera {
-    public static int PosX;
-    public static int PosY;
+    public static float PosX;
+    public static float PosY;
     
     public Camera() {
